+++ tests/micropython/viper_misc.py.exp
@@ -3,15 +3,6 @@
 0
 Ellipsis
 6
+10
 7
 20
-49994955
-1 1
-1 3
-(0, 1, False, True)
-(1, 3)
-[1, 3]
-[1, 3]
-OSError(1,)
-1
-1
