+++ tests/micropython/viper_misc3.py.exp
@@ -1,0 +1,10 @@
+49994955
+1 1
+1 3
+(0, 1, False, True)
+(1, 3)
+[1, 3]
+[1, 3]
+OSError(1,)
+1
+1
